+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Changed
 - Allow for unknown content-disposition types and added a flag to enforce only known.
 
+### Bugfix
+- Import ParseException directly from pyparsing.
+
 ## [1.0.0] - 2022-02-20
 ### Added
 - Initial release